+++ ParkingSpotsSearcher/src/model/GasStation.h
@@ -10,7 +10,7 @@
 
 class GasStation : public Node {
 public:
-    GasStation(long id, float latitude_degrees, float longitude_degrees);
+    GasStation(long id, double latitude_degrees, double longitude_degrees);
 
     NodeType  getType() const;
 };
