+++ ParkingSpotsSearcher/src/model/Node.cpp
@@ -4,7 +4,7 @@
 
 #include "Node.h"
 
-Node::Node(long id, float latitude_degrees, float longitude_degrees) {
+Node::Node(long id, double latitude_degrees, double longitude_degrees) {
     this->id = id;
     this->latitude = latitude_degrees;
     this->longitute = longitude_degrees;
@@ -14,11 +14,11 @@ long Node::getId() const {
     return id;
 }
 
-float Node::getLatitude() const {
+double Node::getLatitude() const {
     return latitude;
 }
 
-float Node::getLongitute() const {
+double Node::getLongitute() const {
     return longitute;
 }
