+++ ParkingSpotsSearcher/src/model/StreetNode.cpp
@@ -4,7 +4,7 @@
 
 #include "StreetNode.h"
 
-StreetNode::StreetNode(long id, float latitude_degrees, float longitude_degrees) : Node(id, latitude_degrees,
+StreetNode::StreetNode(long id, double latitude_degrees, double longitude_degrees) : Node(id, latitude_degrees,
                                                                                         longitude_degrees) {
 
 }
